Fixed type conversion and unused variables warnings

+++ src/Main.cpp
@@ -3,7 +3,7 @@
 int main()
 {
     sf::VideoMode desktop = sf::VideoMode::getDesktopMode();
-    sf::RenderWindow window(sf::VideoMode(BOARD_SQUARE_INIT_SIZE * BOARD_WIDTH, BOARD_SQUARE_INIT_SIZE * BOARD_HEIGHT), "Chess", sf::Style::Titlebar | sf::Style::Resize | sf::Style::Close);
+    sf::RenderWindow window(sf::VideoMode(static_cast<unsigned int>(BOARD_SQUARE_INIT_SIZE * BOARD_WIDTH), static_cast<unsigned int>(BOARD_SQUARE_INIT_SIZE * BOARD_HEIGHT)), "Chess", sf::Style::Titlebar | sf::Style::Resize | sf::Style::Close);
 
     window.setPosition(sf::Vector2i((desktop.width - window.getSize().x) / 2, (desktop.height - window.getSize().y) / 2));
     window.setVerticalSyncEnabled(true);
